Features added (Hashing): Graphemes are now ID'd by their own hash.
* The core idea is to break words up into their graphemes (we don't care
about phonemes--how it is said, but how it is written). Once broken into
graphemes, we convert them into unique hash codes. The hashes are stored
as a list of 32-bit INTEGER values, which take the place of the "word".
The idea is to trim words with letters down into a small set of integers
which can be searched very quickly.